Optional was added, some functionality migrated to service layer from dao

+++ Repository/src/main/java/com/epam/esm/dao/impl/GiftCertificateDAOImpl.java
@@ -2,6 +2,7 @@ package com.epam.esm.dao.impl;
 
 import com.epam.esm.dao.AbstractDAO;
 import com.epam.esm.dao.GiftCertificateDAO;
+import com.epam.esm.dao.TagDAO;
 import com.epam.esm.dateiniso.DateGenerator;
 import com.epam.esm.domain.GiftCertificate;
 import com.epam.esm.domain.Tag;
@@ -14,7 +15,6 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Repository;
 import org.springframework.util.MultiValueMap;
 
-import javax.persistence.NoResultException;
 import javax.persistence.criteria.*;
 import java.util.*;
 
@@ -25,18 +25,19 @@ public class GiftCertificateDAOImpl extends AbstractDAO implements GiftCertifica
 
     private final DateGenerator dateGenerator;
     private final SQLBuilder sqlBuilder;
+    private final TagDAO tagDAO;
 
     @Autowired
-    public GiftCertificateDAOImpl(SessionFactory sessionFactory, DateGenerator dateGenerator, SQLBuilder sqlBuilder) {
+    public GiftCertificateDAOImpl(SessionFactory sessionFactory, DateGenerator dateGenerator, SQLBuilder sqlBuilder, TagDAO tagDAO) {
         super(sessionFactory);
         this.dateGenerator = dateGenerator;
         this.sqlBuilder = sqlBuilder;
+        this.tagDAO = tagDAO;
     }
 
     @Override
-    public GiftCertificate read(long id) {
-        Session session = sessionFactory.getCurrentSession();
-        return session.get(GiftCertificate.class, id);
+    public Optional<GiftCertificate> read(long id) {
+        return Optional.ofNullable(sessionFactory.getCurrentSession().get(GiftCertificate.class, id));
     }
 
     @Override
@@ -63,8 +64,9 @@ public class GiftCertificateDAOImpl extends AbstractDAO implements GiftCertifica
         Session session = sessionFactory.getCurrentSession();
         List<GiftCertificate> giftCertificates = null;
         String tagName = mapOfSearchParams.get(TAG_NAME);
-        Tag tag = getTagByName(tagName);
-        if (Objects.nonNull(tag)) {
+        Optional<Tag> optionalTag = tagDAO.getTagByName(tagName);
+        if (optionalTag.isPresent()) {
+            Tag tag = optionalTag.get();
             String sql = sqlBuilder.createSearchSQL(mapOfSearchParams, getCertificateIdsByTag(tag), page, size);
             giftCertificates = session.createNativeQuery(sql, GiftCertificate.class).getResultList();
         }
@@ -76,12 +78,12 @@ public class GiftCertificateDAOImpl extends AbstractDAO implements GiftCertifica
         Session session = sessionFactory.getCurrentSession();
         Set<Tag> tags = new HashSet<>();
         for (Tag tag : giftCertificate.getTags()) {
-            if (getTagCountByName(tag).equals(0L)) {
+            if (tagDAO.getTagCountByName(tag).equals(0L)) {
                 tags.add(tag);
                 session.save(tag);
             } else {
-                Tag existingTag = getTagByName(tag.getName());
-                tags.add(existingTag);
+                Optional<Tag> optionalExistingTag = tagDAO.getTagByName(tag.getName());
+                optionalExistingTag.ifPresent(tags::add);
             }
         }
         giftCertificate.setTags(tags);
@@ -94,56 +96,23 @@ public class GiftCertificateDAOImpl extends AbstractDAO implements GiftCertifica
     @Override
     public void delete(long id) throws RepositoryException {
         Session session = sessionFactory.getCurrentSession();
-        try {
-            GiftCertificate giftCertificate = getCertificateById(id);
+        Optional<GiftCertificate> optionalGiftCertificate = read(id);
+        if (optionalGiftCertificate.isPresent()){
+            GiftCertificate giftCertificate = optionalGiftCertificate.get();
             session.delete(giftCertificate);
-        } catch (NoResultException e) {
-            throw new RepositoryException();
+        } else {
+            throw new RepositoryException("There is no such id");
         }
     }
 
     @Override
-    public void update(long id, GiftCertificate giftCertificate) throws RepositoryException {
-        Session session = sessionFactory.getCurrentSession();
-        GiftCertificate existingCertificate;
-        try {
-            existingCertificate = getCertificateById(id);
-        } catch (NoResultException e) {
-            throw new RepositoryException();
-        }
-        session.detach(existingCertificate);
-        if (Objects.isNull(giftCertificate.getName())) {
-            giftCertificate.setName(existingCertificate.getName());
-        }
-        if (Objects.isNull(giftCertificate.getDescription())) {
-            giftCertificate.setDescription(existingCertificate.getDescription());
-        }
-        if (Objects.isNull(giftCertificate.getPrice())) {
-            giftCertificate.setPrice(existingCertificate.getPrice());
-        }
-        if (Objects.isNull(giftCertificate.getDuration())) {
-            giftCertificate.setDuration(existingCertificate.getDuration());
-        }
-        giftCertificate.setCreateDate(existingCertificate.getCreateDate());
-        giftCertificate.setLastUpdateDate(dateGenerator.getCurrentDateAsISO());
-        Set<Tag> tags = new HashSet<>();
-        if (Objects.isNull(giftCertificate.getTags()) || giftCertificate.getTags().isEmpty()) {
-            giftCertificate.setTags(existingCertificate.getTags());
-        } else {
-            for (Tag tag : giftCertificate.getTags()) {
-                if (getTagCountByName(tag).equals(0L)) {
-                    tags.add(tag);
-                    session.save(tag);
-                } else {
-                    Tag existingTag = getTagByName(tag.getName());
-                    tags.add(existingTag);
-                    session.update(existingTag);
-                }
-            }
-            giftCertificate.setTags(tags);
-        }
-        giftCertificate.setId(id);
-        session.update(giftCertificate);
+    public void update(long id, GiftCertificate giftCertificate) {
+        sessionFactory.getCurrentSession().update(giftCertificate);
+    }
+
+    @Override
+    public void detachCertificate(GiftCertificate giftCertificate) {
+        sessionFactory.getCurrentSession().detach(giftCertificate);
     }
 
     @Override
@@ -166,39 +135,6 @@ public class GiftCertificateDAOImpl extends AbstractDAO implements GiftCertifica
             query.setMaxResults(size);
         }
         return query.getResultList();
-    }
-
-    private Long getTagCountByName(Tag tag) {
-        CriteriaBuilder criteriaBuilder = sessionFactory.getCurrentSession().getCriteriaBuilder();
-        CriteriaQuery<Long> criteriaQuery = criteriaBuilder.createQuery(Long.class);
-        Root<Tag> root = criteriaQuery.from(Tag.class);
-        criteriaQuery.select(criteriaBuilder.count(root));
-        criteriaQuery.where(criteriaBuilder.equal(root.get(NAME), tag.getName()));
-        return sessionFactory.createEntityManager().createQuery(criteriaQuery).getSingleResult();
-    }
-
-    private Tag getTagByName(String tagName) {
-        Session session = sessionFactory.getCurrentSession();
-        CriteriaBuilder criteriaBuilder = session.getCriteriaBuilder();
-        CriteriaQuery<Tag> criteriaQuery = criteriaBuilder.createQuery(Tag.class);
-        Root<Tag> root = criteriaQuery.from(Tag.class);
-        ParameterExpression<String> parameterExpression = criteriaBuilder.parameter(String.class);
-        criteriaQuery.select(root).where(criteriaBuilder.equal(root.get(NAME), parameterExpression));
-        Query<Tag> query = session.createQuery(criteriaQuery);
-        query.setParameter(parameterExpression, tagName);
-        return query.getSingleResult();
-    }
-
-    private GiftCertificate getCertificateById(long id) {
-        Session session = sessionFactory.getCurrentSession();
-        CriteriaBuilder criteriaBuilder = session.getCriteriaBuilder();
-        CriteriaQuery<GiftCertificate> criteriaQuery = criteriaBuilder.createQuery(GiftCertificate.class);
-        Root<GiftCertificate> root = criteriaQuery.from(GiftCertificate.class);
-        ParameterExpression<Long> parameterExpression = criteriaBuilder.parameter(Long.class);
-        criteriaQuery.select(root).where(criteriaBuilder.equal(root.get(ID), parameterExpression));
-        Query<GiftCertificate> query = session.createQuery(criteriaQuery);
-        query.setParameter(parameterExpression, id);
-        return query.getSingleResult();
     }
 
     private List<Long> getCertificateIdsByTag(Tag tag) {
